Refactor FXIOS-3736 Improve search by using view model
