+++ firefox-ios/Client/Frontend/Browser/Search/SearchListSection.swift
@@ -1,0 +1,15 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/
+
+import Foundation
+
+enum SearchListSection: Int, CaseIterable {
+    case searchSuggestions
+    case firefoxSuggestions
+    case bookmarks
+    case remoteTabs
+    case history
+    case openedTabs
+    case searchHighlights
+}
